RouteProvider 설정 삭제 -> 기존 router, Link 사용으로 복원
