+++ components/keywords/explainBox/index.tsx
@@ -1,10 +1,9 @@
 import { CommonLayoutBox } from '@components/common/commonStyles';
 import icoClose from '@images/ico_close.png';
-import { HOST_URL } from '@public/assets/url';
 import { KeywordCategory } from '@utils/interface/keywords';
 import dynamic from 'next/dynamic';
 import Image from 'next/image';
-import { useRouter } from '@utils/hook/useRouter/useRouter';
+import { useRouter } from 'next/router';
 import { Suspense } from 'react';
 import styled from 'styled-components';
 
@@ -23,7 +22,7 @@ export default function ExplanationComp({
   explain: string;
   keyword: string;
 }) {
-  const { router } = useRouter();
+  const router = useRouter();
 
   return (
     <ExplanationWrapper>
