+++ utils/hook/useKeywordNavigate.tsx
@@ -1,8 +1,8 @@
-import { useRouter } from '@utils/hook/useRouter/useRouter';
+import { useRouter } from 'next/router';
 import { useCallback } from 'react';
 
 export const useKeywordNavigate = () => {
-  const { router } = useRouter();
+  const router = useRouter();
 
   const moveToKeywordDetail = useCallback((id: string) => {
     router.push(`/keywords/${id}`);
