--- utils/hook/useRouter/index.ts
@@ -1,4 +1,0 @@
-export { RouterProvider } from './routerProvider';
-export { useRouter } from './useRouter';
-export { Link } from './link.tsx';
-export type { pageHistory } from './router.types';
