--- utils/hook/useRouter/router.types.ts
@@ -1,4 +1,0 @@
-export type pageHistory = {
-  pageId: string;
-  path: string;
-};
